Extract abstract and real databases to own files
