+++ notifier/database/drivers/__init__.py
@@ -1,0 +1,3 @@
+from notifier.database.drivers.sqlite import SqliteDriver
+
+DatabaseDriver = SqliteDriver
